+++ src/audio_metadata/__init__.py
@@ -5,12 +5,14 @@ from .__about__ import *
 from .api import *
 from .exceptions import *
 from .formats import *
+from .models import *
 
 __all__ = [
 	*__about__.__all__,
 	*api.__all__,
 	*exceptions.__all__,
 	*formats.__all__,
+	*models.__all__,
 ]