+++ src/audio_metadata/formats/__init__.py
@@ -2,7 +2,6 @@ from .flac import *
 from .id3v1 import *
 from .id3v2 import *
 from .id3v2_frames import *
-from .models import *
 from .mp3 import *
 from .tables import *
 from .vorbis import *
@@ -14,7 +13,6 @@ __all__ = [
 	*id3v1.__all__,
 	*id3v2_frames.__all__,
 	*id3v2.__all__,
-	*models.__all__,
 	*mp3.__all__,
 	*tables.__all__,
 	*vorbis.__all__,