+++ src/audio_metadata/formats/id3v2.py
@@ -17,7 +17,6 @@ from bidict import frozenbidict
 from tbm_utils import AttrMapping
 
 from .id3v2_frames import *
-from .models import Tags
 from .tables import (
 	ID3Version,
 	ID3v2FrameIDs,
@@ -26,6 +25,7 @@ from ..exceptions import (
 	InvalidFrame,
 	InvalidHeader,
 )
+from ..models import Tags
 from ..utils import (
 	datareader,
 	decode_synchsafe_int,