+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 * Convert ``LAMEHeader.unwise_settings_used`` to boolean.
 * Handle ID3 frames not valid for ID3 header version.
 	Ignore them and emit a warning to the user.
+* Move models module to top level.
 
 ### Fixed